fix alert

+++ resources/views/task_statuses/index.blade.php
@@ -25,7 +25,7 @@
                         <td>{{ $status->created_at->format('d.m.Y') }}</td>
                         @if(Auth::check())
                         <td>
-                            <a class="text-danger" href="{{ route('task_statuses.destroy', ['task_status' => $status]) }}" data-method="delete" }}>{{ __('taskStatuses.Delete') }}</a>
+                            <a class="text-danger" href="{{ route('task_statuses.destroy', ['task_status' => $status]) }}" data-method="delete" rel="nofollow" data-confirm="{{ __('taskStatuses.Are you sure?') }}">{{ __('taskStatuses.Delete') }}</a>
                             <a href="{{ route('task_statuses.edit', ['task_status' => $status]) }}">{{ __('taskStatuses.Edit') }}</a>
                         </td>
                         @endif
